requirements-dev.txt
-r requirements.txt
coverage
pytest-cov
moto
